Fix bug introduced in r2086 which would give droids trouble avoiding obstacles. Thanks to Troman for pointing out the problem.
git-svn-id: svn+ssh://svn.gna.org/svn/warzone/trunk@2094 4a71c877-e1ca-e34f-864e-861f7616d084

+++ lib/framework/fractions.h
@@ -68,13 +68,13 @@ typedef float FRACT_D;  /* But isn't this is supposed to be double? - Per */
 #define ROUND(x) ((x)>=0 ? (SDWORD)((x) + 0.5) : (SDWORD)((x) - 0.5))
 
 #define MAKEFRACT(x) ((float)(x))
-#define FRACTmul(x,y) ((x)*(y))
-#define FRACTdiv(x,y) ((x)/(y))
+#define FRACTmul(x,y) ((float)(x)*(float)(y))
+#define FRACTdiv(x,y) ((float)(x)/(float)(y))
 
 #define FRACTCONST(a,b) (((float)(a)) / ((float)(b)))
 #define MAKEFRACT_D(x) ((FRACT_D)(x))
-#define FRACTmul_D(x,y) ((x)*(y))
-#define FRACTdiv_D(x,y) ((x)/(y))
+#define FRACTmul_D(x,y) ((double)(x)*(double)(y))
+#define FRACTdiv_D(x,y) ((double)(x)/(double)(y))
 
 #define MAKEINT_D(x) ((SDWORD)(x))
 #define MAKEINT(x) ((SDWORD)(x))
